requirements.txt
streamlit==1.22.0
mysql-connector-python==8.0.32
pandas==1.5.3 